requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
joblib
python-dotenv
httpx
pytest
